+++ changelog.md
@@ -1,5 +1,6 @@
 ## Upcoming
 * Added back Farming features: Crops
+* Fixed Water Fall Damage feature not working properly
 
 ## 5.2.0.0-alpha
 * Added back Mobs features: Equipment, Misc Mobs, Spawning, Zombie Siege